+++ packages/browser/src/consent.ts
@@ -3,7 +3,7 @@ import { find } from './utils'
 import { assignableWindow, navigator } from './utils/globals'
 import { cookieStore, localStore } from './storage'
 import { PersistentStore } from './types'
-import { includes } from './utils/string-utils'
+import { includes } from '@posthog/core'
 
 const OPT_OUT_PREFIX = '__ph_opt_in_out_'