+++ packages/node/src/extensions/error-tracking/index.ts
@@ -4,12 +4,16 @@ import { PostHogBackendClient } from '../../client'
 import { uuidv7 } from '@posthog/core/vendor/uuidv7'
 import { propertiesFromUnknownInput } from './error-conversion'
 import { EventMessage, PostHogOptions } from '../../types'
+import type { Logger } from '@posthog/core'
+import { BucketedRateLimiter } from '@posthog/core/utils'
 
 const SHUTDOWN_TIMEOUT = 2000
 
 export default class ErrorTracking {
   private client: PostHogBackendClient
   private _exceptionAutocaptureEnabled: boolean
+  private _rateLimiter: BucketedRateLimiter<string>
+  private _logger: Logger
 
   static stackParser: StackParser
   static frameModifiers: StackFrameModifierFn[]
@@ -40,9 +44,20 @@ export default class ErrorTracking {
     }
   }
 
-  constructor(client: PostHogBackendClient, options: PostHogOptions) {
+  constructor(client: PostHogBackendClient, options: PostHogOptions, _logger: Logger) {
     this.client = client
     this._exceptionAutocaptureEnabled = options.enableExceptionAutocapture || false
+    this._logger = _logger
+
+    // by default captures ten exceptions before rate limiting by exception type
+    // refills at a rate of one token / 10 second period
+    // e.g. will capture 1 exception rate limited exception every 10 seconds until burst ends
+    this._rateLimiter = new BucketedRateLimiter({
+      refillRate: 1,
+      bucketSize: 10,
+      refillInterval: 10000, // ten seconds in milliseconds
+      _logger: this._logger,
+    })
 
     this.startAutocaptureIfEnabled()
   }
@@ -54,8 +69,19 @@ export default class ErrorTracking {
     }
   }
 
-  private onException(exception: unknown, hint: EventHint): void {
-    void ErrorTracking.buildEventMessage(exception, hint).then((msg) => {
+  private onException(exception: unknown, hint: EventHint): Promise<void> {
+    return ErrorTracking.buildEventMessage(exception, hint).then((msg) => {
+      const exceptionProperties = msg.properties
+      const exceptionType = exceptionProperties?.$exception_list[0].type ?? 'Exception'
+      const isRateLimited = this._rateLimiter.consumeRateLimit(exceptionType)
+
+      if (isRateLimited) {
+        this._logger.info('Skipping exception capture because of client rate limiting.', {
+          exception: exceptionType,
+        })
+        return
+      }
+
       this.client.capture(msg)
     })
   }