+++ packages/node/src/extensions/feature-flags/feature-flags.ts
@@ -1,6 +1,6 @@
 import { FeatureFlagCondition, FlagProperty, FlagPropertyValue, PostHogFeatureFlag, PropertyGroup } from '../../types'
 import type { FeatureFlagValue, JsonType, PostHogFetchOptions, PostHogFetchResponse } from '@posthog/core'
-import { safeSetTimeout } from '@posthog/core'
+import { safeSetTimeout } from '@posthog/core/utils'
 import { hashSHA1 } from './crypto'
 
 const SIXTY_SECONDS = 60 * 1000