+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,14 @@
 # posthog-react-native
 
+## 4.4.2
+
+### Patch Changes
+
+- [#2219](https://github.com/PostHog/posthog-js/pull/2219) [`44d10c4`](https://github.com/PostHog/posthog-js/commit/44d10c46c5378fa046320b7c50bd046eb1e75994) Thanks [@daibhin](https://github.com/daibhin)! - update @posthog/core
+
+- Updated dependencies [[`44d10c4`](https://github.com/PostHog/posthog-js/commit/44d10c46c5378fa046320b7c50bd046eb1e75994)]:
+  - @posthog/core@1.0.1
+
 ## 4.4.1
 
 ### Patch Changes